AutoresolveIds consistency
The GraphArea AutoresolveIds method was clearing _dataIdsCollection and
resetting _dataIdCounter between vertex and edge processing. This often
resulted in GetNextUniqueId returning IDs that were already in use.

Additionally, if the graph was populated by deserialization,
_dataIdsCollection was not populated with the IDs pulled in during
deserialization. So, the next generated ID could be a duplicate. Calling
AutoresolveIds() takes care of that.