Modifications to allow visibility control from styles
There were several places in the code where the Visibility property was
being set as a local value. That interferes with the use of styles to
control visibility. I modified each place where Visibility was set as a
local value to use SetCurrentValue instead.

In the case of VertexControl, a bit more was needed. The sequence of
events to hide and show VertexControl during the layout process was
still calling SetCurrentValue after my style triggers. So, I created the
GraphAreaBase.PositioningComplete attached property and manipulate that
instead. When the layout is in progress and the controls should be
hidden, the attached property is set false and a VertexControl style
trigger can set the property to Hidden. When the layout is complete and
the controls are positioned, the property is set to true and the style
trigger is no longer in effect. Then my style triggers can take over and
properly control the visibility.