Fixed memory leak that was due to use of DependencyPropertyDescriptor.AddValueChanged
The VertexControl class uses
DependencyPropertyDescriptor.AddValueChanged to receive change
notifications on GraphAreaBase.X and GraphAreaBase.Y attached
properties. It was not removing the value change monitoring and had a
memory leak as a result. Modified the control position change handling
to use bindings instead of DependencyPropertyDescriptor.AddValueChanged.
This eliminates strong references from the property descriptor to the
vertex controls.