Modified drag and click handling and implemented a Click event on VertexControl. The Click event should have similar behavior to that of Button controls, which do not fire the Click event until the button is released. Here, the Click event is only wanted if the user doesn't drag the control.
Several changes are made to DragBehavior, mainly to reduce conflicts between VertexControl and DragBehavior trying to handle the same events and conflicting with each other. For example, prior to these changes, the VertexControl was not getting the MouseUp events it was expecting.

VertexControl MouseDoubleClick handler no longer marks the event handled because it prevents parent controls from receiving the bubbling event. Marking the other mouse events handled might be undesirable as well because notifying the RootArea of the events doesn't necessarily mean other elements of the visual tree shouldn't get a chance at the event. The corresponding RootArea events are fine, but they are different events and bypass the rest of the bubbling that the events on the VertexControl would continue through. In my opinion, the event handlers of the VertexControl should not mark the events handled, but I have only changed that for MouseDoubleClick at this time.