Update GraphArea.cs (#95)
* Update GraphArea.cs

Have noticed that sometimes during DeSerialization a failure for Source or Target to be updated properly will result in an uncaught exception to the Application Dispatcher.  It is best to test for a valid Source and Target Property value before attempting to retrieve a mapped value.

* Update GraphArea.cs

Testing revealed an issue.